MSI-1486: create ui on stock edit page and source edit page

+++ app/code/Magento/InventoryConfiguration/Model/ConfigurationOptions/GetMinSaleQtyConfigurationValue.php
@@ -55,6 +55,6 @@ class GetMinSaleQtyConfigurationValue implements GetMinSaleQtyConfigurationValue
      */
     public function forGlobal(): float
     {
-        return (int)$this->getConfigurationValue->execute(StockItemConfigurationInterface::MIN_SALE_QTY);
+        return (float)$this->getConfigurationValue->execute(StockItemConfigurationInterface::MIN_SALE_QTY);
     }
 }
